Refactor Python server handling: Moved Python server start/stop functions to a new module and updated import paths. Removed outdated test files for Python server functionality, streamlining the codebase and enhancing maintainability.
