+++ Frontend/src/electron/python/startAndStopPython.ts
@@ -1,231 +1,21 @@
-import { app, dialog, shell } from "electron";
-import { spawn, ChildProcess, execSync, SpawnOptions } from "child_process";
+import { app } from "electron";
+import { spawn, ChildProcess, SpawnOptions } from "child_process";
 import path from "path";
-import { isDev } from "./util.js";
-import { updateLoadingStatus } from "./loadingWindow.js";
+import { isDev } from "../util.js";
+import { updateLoadingStatus } from "../loadingWindow.js";
 import fs from "fs";
 import log from "electron-log";
 import ffmpegStatic from "ffmpeg-static";
-import { generateSecret } from "./authentication/secret.js";
-import { getSecret } from "./authentication/devApi.js";
-
+import { generateSecret } from "../authentication/secret.js";
+import { getSecret } from "../authentication/devApi.js";
+import { ensurePythonAndVenv } from "./ensurePythonAndVenv.js";
+import { extractFromAsar } from "./extractFromAsar.js";
 log.transports.file.level = "info";
 log.transports.file.resolvePathFn = () =>
   path.join(app.getPath("userData"), "logs/main.log");
 
 let pythonProcess: ChildProcess | null = null;
 
-async function runWithPrivileges(commands: string | string[]): Promise<void> {
-  if (process.platform !== "linux") return;
-
-  const commandArray = Array.isArray(commands) ? commands : [commands];
-
-  try {
-    // Try without privileges first
-    for (const cmd of commandArray) {
-      execSync(cmd);
-    }
-  } catch {
-    log.info("Failed to run commands, requesting privileges..., ");
-
-    const response = await dialog.showMessageBox({
-      type: "question",
-      buttons: ["Grant Privileges", "Cancel"],
-      defaultId: 0,
-      title: "Administrator Privileges Required",
-      message:
-        "Creating the Python environment requires administrator privileges.",
-      detail:
-        "This is needed to install required system dependencies and create the virtual environment. This will only be needed once.",
-    });
-
-    if (response.response === 0) {
-      try {
-        // Combine all commands with && to run them in sequence
-        const combinedCommand = commandArray.join(" && ");
-        execSync(`pkexec sh -c '${combinedCommand}'`);
-      } catch (error) {
-        log.error("Failed to run commands with privileges", error);
-        throw new Error("Failed to run commands with elevated privileges");
-      }
-    } else {
-      throw new Error(
-        "User declined to grant administrator privileges. Cannot continue."
-      );
-    }
-  }
-}
-
-async function ensurePythonAndVenv(backendPath: string) {
-  const venvPath = path.join(backendPath, "venv");
-  const pythonCommands =
-    process.platform === "win32"
-      ? ["python3.10", "py -3.10", "python"]
-      : process.platform === "darwin"
-      ? ["/opt/homebrew/bin/python3.10", "python3.10", "python3"]
-      : ["python3.10", "python3"];
-
-  let pythonCommand: string | null = null;
-  let pythonVersion: string | null = null;
-
-  for (const cmd of pythonCommands) {
-    try {
-      log.info(`Trying Python command: ${cmd}`);
-      const version = execSync(`${cmd} --version`).toString().trim();
-      log.info(`Version output: ${version}`);
-      if (version.includes("3.10")) {
-        pythonCommand = cmd;
-        pythonVersion = version;
-        log.info(`Found valid Python command: ${cmd} with version ${version}`);
-        break;
-      }
-    } catch (error: unknown) {
-      if (error instanceof Error) {
-        log.info(`Failed to execute ${cmd}: ${error.message}`);
-      }
-      continue;
-    }
-  }
-
-  if (!pythonCommand) {
-    log.error("Python 3.10 is not installed or not in PATH");
-    const response = await dialog.showMessageBox({
-      type: "question",
-      buttons: ["Install Python 3.10", "Cancel"],
-      defaultId: 0,
-      title: "Python 3.10 Required",
-      message: "Python 3.10 is required but not found on your system.",
-      detail:
-        "Would you like to open the Python download page to install Python 3.10?",
-    });
-
-    if (response.response === 0) {
-      // Open Python download page
-      await shell.openExternal(
-        "https://www.python.org/downloads/release/python-31010/"
-      );
-      throw new Error(
-        "Please restart the application after installing Python 3.10"
-      );
-    } else {
-      throw new Error(
-        "Python 3.10 is required to run this application. Installation was cancelled."
-      );
-    }
-  }
-
-  log.info(`Using ${pythonVersion}`);
-
-  const venvPython =
-    process.platform === "win32"
-      ? path.join(venvPath, "Scripts", "python.exe")
-      : path.join(venvPath, "bin", "python");
-
-  if (!fs.existsSync(venvPath)) {
-    log.info("Creating virtual environment with Python 3.10...");
-
-    if (process.platform === "linux") {
-      try {
-        // Run all commands with a single privilege prompt
-        await runWithPrivileges([
-          "apt-get update && apt-get install -y python3-venv python3-dev build-essential",
-          `${pythonCommand} -m venv "${venvPath}"`,
-          `chown -R ${process.env.USER}:${process.env.USER} "${venvPath}"`,
-        ]);
-
-        log.info("Virtual environment created successfully");
-      } catch (error: unknown) {
-        if (error instanceof Error) {
-          log.error("Failed to create virtual environment", error);
-          throw error;
-        }
-        throw new Error("Unknown error while creating virtual environment");
-      }
-    } else {
-      // Original code for non-Linux systems
-      try {
-        execSync(`${pythonCommand} -m venv "${venvPath}"`);
-        log.info("Virtual environment created successfully");
-      } catch (error: unknown) {
-        if (error instanceof Error) {
-          log.error("Failed to create virtual environment", error);
-          throw new Error("Failed to create virtual environment");
-        } else {
-          log.error("Unknown error in ensurePythonAndVenv", error);
-          throw new Error("Unknown error in ensurePythonAndVenv");
-        }
-      }
-    }
-  }
-
-  try {
-    execSync(`"${venvPython}" -m pip install --upgrade pip`);
-    log.info("Pip upgraded successfully");
-  } catch (error) {
-    log.error("Failed to upgrade pip", error);
-    throw new Error("Failed to upgrade pip");
-  }
-
-  // Add check for NVIDIA GPU
-  let hasNvidiaGpu = false;
-  try {
-    if (process.platform === "linux") {
-      execSync("nvidia-smi");
-      hasNvidiaGpu = true;
-    } else if (process.platform === "win32") {
-      execSync("nvidia-smi");
-      hasNvidiaGpu = true;
-    } else if (process.platform === "darwin") {
-      // MacOS doesn't support CUDA
-      hasNvidiaGpu = false;
-    }
-  } catch {
-    log.info("No NVIDIA GPU detected, will use CPU-only packages");
-    hasNvidiaGpu = false;
-  }
-
-  // Set environment variable for the Python process
-  process.env.USE_CUDA = hasNvidiaGpu ? "1" : "0";
-
-  return { venvPython, hasNvidiaGpu };
-}
-
-function extractFromAsar(sourcePath: string, destPath: string) {
-  log.info(`Extracting from ${sourcePath} to ${destPath}`);
-  try {
-    if (!fs.existsSync(sourcePath)) {
-      throw new Error(`Source path does not exist: ${sourcePath}`);
-    }
-    if (!fs.existsSync(destPath)) {
-      log.info(`Creating directory: ${destPath}`);
-      fs.mkdirSync(destPath, { recursive: true });
-    }
-
-    const files = fs.readdirSync(sourcePath);
-    log.info(`Files in source: ${files.join(", ")}`);
-    files.forEach((file) => {
-      const fullSourcePath = path.join(sourcePath, file);
-      const fullDestPath = path.join(destPath, file);
-
-      if (fs.statSync(fullSourcePath).isDirectory()) {
-        log.info(`Extracting directory: ${file}`);
-        extractFromAsar(fullSourcePath, fullDestPath);
-      } else {
-        log.info(`Copying file: ${file}`);
-        fs.copyFileSync(fullSourcePath, fullDestPath);
-      }
-    });
-    log.info(`Extraction completed for ${sourcePath}`);
-  } catch (error: unknown) {
-    if (error instanceof Error) {
-      log.error(`Error in extractFromAsar: ${error.message}`);
-      log.error(`Stack trace: ${error.stack}`);
-    } else {
-      log.error(`Unknown error in extractFromAsar: ${error}`);
-    }
-    throw error;
-  }
-}
 
 export async function startPythonServer() {
   log.info("Application starting...");
